[refactor] Modularize db.js into separate models and make template for controllers
