+++ server/controllers/itinerary.controller.js
@@ -1,4 +1,4 @@
-//var savedActivitiesModel = require('../models').savedActivitiesModel;
+var SavedActivities = require('../db').SavedActivities;
 
 var itineraryController = {};
 
@@ -9,7 +9,15 @@ itineraryController.GET = function(req, res) {
 
 itineraryController.POST = function(req, res) {
   console.log('inside itineraryController.POST');
-  res.status(200).send();
+  SavedActivities.sync().then(function() {
+    return SavedActivities.create({
+      name: 'testname',
+      picture: 'testpic',
+      description: 'testdesc',
+      address: 'testaddress'
+    });
+  });
+  res.status(201).send();
 };
 
 module.exports = {
