+++ server/models/index.js
@@ -1,0 +1,9 @@
+var PossibleActivities = require('../models/possibleActivities');
+var SavedActivities = require('../models/savedActivities');
+var Trip = require('../models/trip');
+
+module.exports = {
+  PossibleActivities: PossibleActivities,
+  SavedActivities: SavedActivities,
+  Trip: Trip
+}
